Move TaskStep to separate file and remove from Task.swift
